[redesign] Remove LpShare contract
The LpShare contract implementation reveal
to be too complex and potentially subject
to economical exploits/gaming

changing the design to use a unique RevenueShare
contract. And build a revenue to LP buy flow
in order to feed the Lp staking contract

The idea is that you deposit LP tokens, and
over time you will get more Lp tokens. The
same way the $METRIC staking contract work.
This also allows minting a new token for
staked LP positions that can be used
for governance